DifferentiableFrankWolfe

Documentation for DifferentiableFrankWolfe.

DifferentiableFrankWolfe.ConditionsFW

DifferentiableFrankWolfe.DiffFW

DifferentiableFrankWolfe.DiffFW

DifferentiableFrankWolfe.ForwardFW

DifferentiableFrankWolfe.simplex_projection

DifferentiableFrankWolfe.simplex_projection_and_support

DifferentiableFrankWolfe.ConditionsFW — Type

ConditionsFW{F,G,M}

Differentiable optimality conditions for DiffFW, which rely on a custom sparse_argmax implementation.

DifferentiableFrankWolfe.DiffFW — Type

DiffFW{F,G,M,A,I}

Callable parametrized wrapper for the Frank-Wolfe algorithm θ -> argmin_{x ∈ C} f(x, θ), which can be differentiated implicitly wrt θ.

Reference: https://arxiv.org/abs/2105.15183 (especially section 2 and the end of appendix A).

Fields

f::F: function f(x, θ) to minimize wrt x

f_grad1::G: gradient ∇ₓf(x, θ) of f wrt x

lmo::M: linear minimization oracle θ -> argmin_{x ∈ C} θᵀx, implicitly defines the convex set C

alg::A: Frank-Wolfe variante used, defaults to away_frank_wolfe

implicit::I: implicit function constructed from the previous fields

DifferentiableFrankWolfe.DiffFW — Method

dfw(θ; frank_wolfe_kwargs)

Apply the Frank-Wolfe algorithm to θ with settings defined by frank_wolfe_kwargs.

DifferentiableFrankWolfe.ForwardFW — Type

ForwardFW{F,G,M,A}

Underlying solver for DiffFW, which relies on a variant of Frank-Wolfe.

DifferentiableFrankWolfe.simplex_projection — Method

simplex_projection(z)

Compute the Euclidean projection of the vector z onto the probability simplex.

Reference: https://arxiv.org/abs/1602.02068.

DifferentiableFrankWolfe.simplex_projection_and_support — Method

simplex_projection_and_support(z)

Compute the Euclidean projection p of z on the probability simplex and the indicators s of its support.

Reference: https://arxiv.org/abs/1602.02068.